fix: use tailwind
